+++ application/state.py
@@ -1,5 +1,13 @@
 
 
+#Application state on module
+
+current_namespace=""
+selected_pod=""
+selected_node="all"
+searchString=""
+
+
 class State(object):
 
     def __init__(self):