+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Added red to some error/exception lines in Output-window.
 - Added "all worker nodes" to Nodes-window.
 - Added pretty printing to Pods-window.
+- Added pod count to Pods-window.
 
 ## 0.2